gpc/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
gspread==5.7.2
gunicorn==20.1.0
lxml==4.9.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
sqlalchemy==2.0.6
